# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API DE O BI COOK",
        "description": "Documentation des API de O BI COOK",
        "contact": {
            "name": "Souleymane TONDE",
            "email": "tondesoulco@gmail.com"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Enregistrer un nouvel utilisateur",
                "description": "Crée un nouvel utilisateur et envoie un email de vérification OTP.",
                "operationId": "735073bfc89fc9fb460bac5e1e48c690",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "description": "Prénom de l'utilisateur",
                                        "type": "string",
                                        "example": "Jean"
                                    },
                                    "last_name": {
                                        "description": "Nom de famille de l'utilisateur",
                                        "type": "string",
                                        "example": "Dupont"
                                    },
                                    "phone_number": {
                                        "description": "Numéro de téléphone de l'utilisateur",
                                        "type": "string",
                                        "example": "+2250102030405"
                                    },
                                    "email": {
                                        "description": "Adresse email de l'utilisateur",
                                        "type": "string",
                                        "example": "jean.dupont@example.com"
                                    },
                                    "password": {
                                        "description": "Mot de passe de l'utilisateur",
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 201
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Un email de vérification a été envoyé à votre adresse email"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 422
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Les données de l'utilisateur sont invalides"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/verify-email": {
            "get": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Vérifier l'adresse email de l'utilisateur",
                "description": "Endpoint pour vérifier l'adresse email de l'utilisateur via un code OTP.",
                "operationId": "b94a790183acde18b139e635c8d6170a",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "L'adresse email de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "email"
                        }
                    },
                    {
                        "name": "otp",
                        "in": "query",
                        "description": "Le code OTP envoyé à l'adresse email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Votre adresse email a été vérifiée"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 422
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Le code de vérification est invalide"
                                        },
                                        "errors": {
                                            "properties": {
                                                "otp": {
                                                    "type": "string",
                                                    "example": "Le code de vérification est invalide"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 401
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Non autorisé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Utilisez le token Bearer pour accéder aux API protégées",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}